requirements.txt
fastapi
uvicorn[standard]
pytesseract
pdfplumber
pillow
